adding confirm to delete the data + changing font size for inputs element in creditsTable and adding type number

+++ deleteData.js
@@ -2,7 +2,13 @@
 
 function deleteDb()
 {
-    localStorage.clear()
+    if (confirm("Est ce que vous voulez effacer vos données")) {
+        //delete 
+        localStorage.clear()
+      } else {
+        //Nothing happened
+      }
+    
 }
 
 export default  deleteDb
